backend/requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Flask==3.0.0
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==24.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
Werkzeug==3.0.1
